Refactor: 랭킹조회 score double에서 int로 변경

+++ backend/raon/src/main/java/com/arch/raon/domain/grammar/dto/redis/GrammarMyRankRedisDTO.java
@@ -7,21 +7,21 @@ import org.springframework.data.redis.core.ZSetOperations;
 @Getter
 public class GrammarMyRankRedisDTO {
     private String ranker;
-    private double score;
+    private int score;
 
     public GrammarMyRankRedisDTO() {
         super();
     }
 
     @Builder
-    public GrammarMyRankRedisDTO(String ranker, double score) {
+    public GrammarMyRankRedisDTO(String ranker, int score) {
         this.ranker = ranker;
         this.score = score;
     }
 
     public static GrammarMyRankRedisDTO convertToGrammarMyRankRedisDTO(ZSetOperations.TypedTuple<String> typedTuple){
         String nickname = typedTuple.getValue();
-        double score = typedTuple.getScore();
+        int score = typedTuple.getScore().intValue();
         return new GrammarMyRankRedisDTO(nickname, score);
     }
